+++ allpress/exceptions.py
@@ -13,3 +13,13 @@ class TranslationError:
 
     def __init__(self, msg):
         super().__init__(msg)
+
+class NoSuchColumnError(Exception):
+
+    def __init__(self, msg):
+        super().__init__(msg)
+
+class URINotProvidedError(Exception):
+
+    def __init__(self, msg):
+        super().__init__(msg)