+++ allpress/settings.py
@@ -8,3 +8,14 @@ POSTGRESQL_PASSWORD        = 'test0000'
 
 URL_REGEX = '((http|https):\/\/)?((www|ww\d|www\d)\.)?(?=.{5,255})([\w-]{2,63}\.)+\w{2,63}(\/[\w\-._~:?#@!$&\'\(\)*+,;%=]+)*'
 HREF_PARSE_REGEX = '(?<=<a\shref=([\'"]))([\w\-._~:?#@!$&/\'\(\)*+,;%=]+)\1'
+
+CONTINENTS = [
+    'Africa',
+    'Asia',
+    'Europe',
+    'North America',
+    'South America',
+    'Oceania'
+]
+
+NEWS_SOURCE_CATALOG_FILE = '/home/hassan2022cbtest/allpress/sources.csv'